+++ src/main/java/com/java/project/paragon/ParagonPOJO.java
@@ -14,7 +14,7 @@ public class ParagonPOJO {
     private String miasto;
     private String kodPocztowy;
     private Integer kebab;
-    private Integer turek;
+    private Long turek;
     private Double suma;
 
 }
